BPE26-ref: extract the fullscreen button section to dedicated file
