+++ src/features/ui/index.ts
@@ -1,0 +1,1 @@
+export { createFullscreenButtonSection } from "./fullscreen-button";
